Refactor the package structure for open-api doc generation
