+++ openapi-extension/src/main/java/io/ballerina/openapi/extension/doc/gen/SingleFileOpenApiDocGenerator.java
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-package io.ballerina.openapi.extension.doc;
+package io.ballerina.openapi.extension.doc.gen;
 
 import io.ballerina.openapi.extension.Constants;
-import io.ballerina.openapi.extension.doc.gen.AbstractOpenApiDocGenerator;
 import io.ballerina.projects.ProjectKind;
 
 import java.nio.file.Files;
